Prepare ForexService for method extraction

+++ src/main/kotlin/com/adolfoeloy/taxtracker/forex/ForexService.kt
@@ -30,5 +30,5 @@ interface ForexService {
     fun getForexRateFor(
         currencyTicker: String,
         date: LocalDate
-    ): BigDecimal
+    ): Int
 }
